Bandaid workaround of the culling issue
The nature of TER makes TER invisible when it is fully blocked, i.e.
culling. However, sometimes we do not want the culling, for example when
we want to hide the slide projector behind a wall.
This commit serves as a bandaid workaround to forcibly "turn off" the
culling process. Performance penalty are expected.